Move some files in json_parser folder into a different subfolders
