+++ src/client/CMakeLists.txt
@@ -12,18 +12,18 @@ target_sources(
   PRIVATE "network/http_curl.c"
           "api/json_parser/common.c"
           "api/json_parser/json_utils.c"
-          "api/json_parser/unlock_blocks.c"
-          "api/json_parser/inputs.c"
-          "api/json_parser/feat_blocks.c"
-          "api/json_parser/native_tokens.c"
-          "api/json_parser/output_alias.c"
-          "api/json_parser/output_basic.c"
-          "api/json_parser/output_foundry.c"
-          "api/json_parser/output_nft.c"
-          "api/json_parser/outputs.c"
-          "api/json_parser/unlock_conditions.c"
           "api/json_parser/message.c"
-          "api/json_parser/payloads.c"
+          "api/json_parser/unlock_blocks.c"
+          "api/json_parser/inputs/inputs.c"
+          "api/json_parser/outputs/feat_blocks.c"
+          "api/json_parser/outputs/native_tokens.c"
+          "api/json_parser/outputs/output_alias.c"
+          "api/json_parser/outputs/output_basic.c"
+          "api/json_parser/outputs/output_foundry.c"
+          "api/json_parser/outputs/output_nft.c"
+          "api/json_parser/outputs/outputs.c"
+          "api/json_parser/outputs/unlock_conditions.c"
+          "api/json_parser/payloads/payloads.c"
           "api/restful/get_node_info.c"
           "api/restful/response_error.c"
           "api/restful/get_tips.c"
@@ -43,18 +43,18 @@ target_sources(
          "api/json_parser/json_utils.h"
          "api/json_parser/json_keys.h"
          "api/json_parser/common.h"
+          "api/json_parser/message.h"
          "api/json_parser/unlock_blocks.h"
-         "api/json_parser/inputs.h"
-         "api/json_parser/feat_blocks.h"
-         "api/json_parser/native_tokens.h"
-         "api/json_parser/output_alias.h"
-         "api/json_parser/output_basic.h"
-         "api/json_parser/output_foundry.h"
-         "api/json_parser/output_nft.h"
-         "api/json_parser/outputs.h"
-         "api/json_parser/unlock_conditions.h"
-         "api/json_parser/message.h"
-         "api/json_parser/payloads.h"
+         "api/json_parser/inputs/inputs.h"
+         "api/json_parser/outputs/feat_blocks.h"
+         "api/json_parser/outputs/native_tokens.h"
+         "api/json_parser/outputs/output_alias.h"
+         "api/json_parser/outputs/output_basic.h"
+         "api/json_parser/outputs/output_foundry.h"
+         "api/json_parser/outputs/output_nft.h"
+         "api/json_parser/outputs/outputs.h"
+         "api/json_parser/outputs/unlock_conditions.h"
+         "api/json_parser/payloads/payloads.h"
          "api/restful/get_node_info.h"
          "api/restful/response_error.h"
          "api/restful/get_tips.h"
