+++ src/client/api/json_parser/outputs/feat_blocks.c
@@ -1,7 +1,7 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#include "client/api/json_parser/feat_blocks.h"
+#include "client/api/json_parser/outputs/feat_blocks.h"
 #include "client/api/json_parser/common.h"
 #include "core/utils/macros.h"
 #include "utlist.h"
