+++ src/client/api/json_parser/outputs/feat_blocks.h
@@ -1,8 +1,8 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef __CLIENT_API_JSON_PARSER_FEAT_BLOCKS_H__
-#define __CLIENT_API_JSON_PARSER_FEAT_BLOCKS_H__
+#ifndef __CLIENT_API_JSON_PARSER_OUTPUTS_FEAT_BLOCKS_H__
+#define __CLIENT_API_JSON_PARSER_OUTPUTS_FEAT_BLOCKS_H__
 
 #include "client/api/json_parser/json_utils.h"
 #include "core/models/outputs/feat_blocks.h"
@@ -69,4 +69,4 @@ cJSON *json_feat_blocks_serialize(feat_blk_list_t *feat_blocks);
 }
 #endif
 
-#endif  // __CLIENT_API_JSON_PARSER_FEAT_BLOCKS_H__
+#endif  // __CLIENT_API_JSON_PARSER_OUTPUTS_FEAT_BLOCKS_H__
