+++ src/client/api/json_parser/outputs/unlock_conditions.h
@@ -1,8 +1,8 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef __CLIENT_API_JSON_PARSER_UNLOCK_CONDITIONS_H__
-#define __CLIENT_API_JSON_PARSER_UNLOCK_CONDITIONS_H__
+#ifndef __CLIENT_API_JSON_PARSER_OUTPUTS_UNLOCK_CONDITIONS_H__
+#define __CLIENT_API_JSON_PARSER_OUTPUTS_UNLOCK_CONDITIONS_H__
 
 #include "client/api/json_parser/json_utils.h"
 #include "core/models/outputs/unlock_conditions.h"
@@ -86,4 +86,4 @@ cJSON *json_cond_blk_list_serialize(cond_blk_list_t *blk_list);
 }
 #endif
 
-#endif  // __CLIENT_API_JSON_PARSER_UNLOCK_CONDITIONS_H__
+#endif  // __CLIENT_API_JSON_PARSER_OUTPUTS_UNLOCK_CONDITIONS_H__
