+++ tests/client/api_json_parser/test_output_alias.c
@@ -3,9 +3,8 @@
 
 #include <stdio.h>
 
-#include "client/api/json_parser/output_alias.h"
+#include "client/api/json_parser/outputs/output_alias.h"
 #include "core/models/outputs/output_alias.h"
-#include "core/models/outputs/outputs.h"
 #include "core/utils/macros.h"
 #include "unity/unity.h"
