+++ tests/client/api_json_parser/test_output_foundry.c
@@ -3,7 +3,7 @@
 
 #include <stdio.h>
 
-#include "client/api/json_parser/output_foundry.h"
+#include "client/api/json_parser/outputs/output_foundry.h"
 #include "core/models/outputs/output_foundry.h"
 #include "core/models/outputs/outputs.h"
 #include "core/utils/macros.h"
