+++ tests/client/api_json_parser/test_unlock_conditions.c
@@ -3,7 +3,7 @@
 
 #include <stdio.h>
 
-#include "client/api/json_parser/unlock_conditions.h"
+#include "client/api/json_parser/outputs/unlock_conditions.h"
 #include "core/utils/macros.h"
 #include "unity/unity.h"
